requirements.txt
discord.py
python-dotenv
requests
youtube-dl
validators
pynacl
ffmpeg
asyncio
openai